+++ src/CoolPropLib.cpp
@@ -781,7 +781,7 @@ EXPORT_CODE void CONVENTION AbstractState_set_binary_interaction_double(const lo
     }
 }
 
-EXPORT_CODE void CONVENTION  AbstractState_set_fluid_parameter_double(const long handle, const size_t i, const char* parameter, const double value , long *errcode, char *message_buffer, const long buffer_length) {
+EXPORT_CODE void CONVENTION  AbstractState_set_fluid_parameter_double(const long handle, const long i, const char* parameter, const double value , long *errcode, char *message_buffer, const long buffer_length) {
     *errcode = 0;
     try {
         shared_ptr<CoolProp::AbstractState> &AS = handle_manager.get(handle);
